+++ assets/css/fonts.css
@@ -1,0 +1,7 @@
+@font-face {
+    font-family: 'Space Grotesk';
+    src: url('~/assets/fonts/SpaceGrotesk-VariableFont_wght.ttf') format('truetype');
+    font-weight: 100 900; /* since it's a variable font */
+    font-style: normal;
+    font-display: swap;
+}
